refactor(connect-scope): extract utilities and SpanDetailPanel from TraceViewer
Move pure functions (formatting, OTLP parsing, trace-building) into
dedicated utility modules and shared types into types/index.ts. Replace
duplicated span detail templates in waterfall and flamegraph views with
a reusable SpanDetailPanel component. Reduces TraceViewer from 820 to
579 lines with no behavioral changes.
