+++ extensions/connect-scope/src/utils/formatting.ts
@@ -1,0 +1,18 @@
+export function formatTime(ms: number): string {
+  return new Date(ms).toLocaleString();
+}
+
+export function formatDuration(ms: number): string {
+  if (ms >= 1000) return (ms / 1000).toFixed(2) + " s";
+  return ms.toFixed(1) + " ms";
+}
+
+export function timeAgo(isoString: string): string {
+  const secs = Math.floor((Date.now() - new Date(isoString).getTime()) / 1000);
+  if (secs < 60) return "just now";
+  const mins = Math.floor(secs / 60);
+  if (mins < 60) return `${mins}m ago`;
+  const hrs = Math.floor(mins / 60);
+  if (hrs < 24) return `${hrs}h ago`;
+  return `${Math.floor(hrs / 24)}d ago`;
+}
